restructured and refactored sudoku components
